# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: VitalMatch API - Connecting hospitals with blood donors in Kenya
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /donors/me:
    get:
      operationId: getMyDonorProfile
      tags: [donors]
      summary: Get current donor profile
      responses:
        "200":
          description: Donor profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DonorProfile"
    patch:
      operationId: updateMyDonorProfile
      tags: [donors]
      summary: Update donor profile and location
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDonorBody"
      responses:
        "200":
          description: Updated donor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DonorProfile"
...
  /donors/me/token:
    get:
      operationId: getDonorToken
      tags: [donors]
      summary: Get or generate QR token for accepted appeal
      responses:
        "200":
          description: Token data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DonorToken"
        "404":
          description: No active accepted appeal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /gemini/conversations:
    get:
      operationId: listGeminiConversations
      tags: [gemini]
      summary: List all conversations
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GeminiConversation"
    post:
      operationId: createGeminiConversation
      tags: [gemini]
      summary: Create a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGeminiConversationBody"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeminiConversation"
  /gemini/conversations/{id}:
    get:
      operationId: getGeminiConversation
      tags: [gemini]
      summary: Get conversation with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeminiConversationWithMessages"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeminiError"
    delete:
      operationId: deleteGeminiConversation
      tags: [gemini]
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeminiError"
  /gemini/conversations/{id}/messages:
    get:
      operationId: listGeminiMessages
      tags: [gemini]
      summary: List messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GeminiMessage"
    post:
      operationId: sendGeminiMessage
      tags: [gemini]
      summary: Send a message and receive an AI response (SSE stream)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendGeminiMessageBody"
      responses:
        "200":
          description: SSE stream of assistant response chunks
          content:
            text/event-stream: {}
  /gemini/generate-image:
    post:
      operationId: generateGeminiImage
      tags: [gemini]
      summary: Generate an image from a text prompt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateGeminiImageBody"
      responses:
        "200":
          description: Generated image data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateGeminiImageResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    DonorProfile:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        bloodType:
          type: string
        dateOfBirth:
          type: string
        lastDonationDate:
          type: ["string", "null"]
          format: date-time
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        isEligible:
          type: boolean
        totalDonations:
          type: integer
        badges:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - name
        - email
        - phone
        - bloodType
        - dateOfBirth
        - lastDonationDate
        - latitude
        - longitude
        - isEligible
        - totalDonations
        - badges
        - createdAt
    UpdateDonorBody:
      type: object
      properties:
        phone:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        lastDonationDate:
          type: string
...
    DonorToken:
      type: object
      properties:
        token:
          type: string
        donorId:
          type: integer
        hospitalId:
          type: integer
        appealId:
          type: integer
        expiresAt:
          type: string
          format: date-time
        qrData:
          type: string
      required:
        - token
        - donorId
        - hospitalId
        - appealId
        - expiresAt
        - qrData
...
    GeminiConversation:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - createdAt
    GeminiMessage:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - role
        - content
        - createdAt
    CreateGeminiConversationBody:
      type: object
      properties:
        title:
          type: string
      required:
        - title
    SendGeminiMessageBody:
      type: object
      properties:
        content:
          type: string
      required:
        - content
    GeminiConversationWithMessages:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/GeminiMessage"
      required:
        - id
        - title
        - createdAt
        - messages
    GenerateGeminiImageBody:
      type: object
      properties:
        prompt:
          type: string
      required:
        - prompt
    GenerateGeminiImageResponse:
      type: object
      properties:
        b64_json:
          type: string
        mimeType:
          type: string
      required:
        - b64_json
        - mimeType
    GeminiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error